zng_wgt_tooltip

Function access_tooltip_anchor

pub fn access_tooltip_anchor(
    child: impl UiNode,
    mode: impl IntoVar<AnchorMode>,
) -> impl UiNode

P Set the position of the tip widgets opened for the widget or its descendants without cursor interaction.

This position is used instead of tooltip_anchor when the tooltip is shown by commands such as ACCESS.show_tooltip and the cursor is not over the widget.

This property sets the ACCESS_TOOLTIP_ANCHOR_VAR.