zng_wgt_tooltip

Function access_tooltip_duration

pub fn access_tooltip_duration(
    child: impl UiNode,
    duration: impl IntoVar<Duration>,
) -> impl UiNode

P Sets the maximum duration a tooltip stays open on the widget or descendants when it is opened without cursor interaction.

This duration is used instead of tooltip_duration when the tooltip is shown by commands such as ACCESS.show_tooltip and the cursor is not over the widget.

Zero means until ACCESS.hide_tooltip, is 5 seconds by default.

This property sets the ACCESS_TOOLTIP_DURATION_VAR.