Function zng_wgt_tooltip::tooltip

pub fn tooltip(child: impl UiNode, tip: impl UiNode) -> impl UiNode

P Widget tooltip.

Any other widget can be used as tooltip, the recommended widget is the Tip! container, it provides the tooltip style. Note that if the tip node is not a widget even after initializing it will not be shown.

This property can be configured by tooltip_anchor, tooltip_delay, tooltip_interval and tooltip_duration.

This tooltip only opens if the widget is enabled, see disabled_tooltip for a tooltip that only shows when the widget is disabled.