Function zng_wgt_tooltip::tooltip_duration

pub fn tooltip_duration(
    child: impl UiNode,
    duration: impl IntoVar<Duration>,
) -> impl UiNode

P Sets the maximum duration a tooltip stays open on the widget or descendants.

Note that the tooltip closes at the moment the cursor leaves the widget, this duration defines the time the tooltip is closed even if the cursor is still hovering the widget. This duration is not used if the tooltip is opened without cursor interaction, in that case the access_tooltip_duration is used.

Zero means indefinitely, is zero by default.

This property sets the TOOLTIP_DURATION_VAR.