zng_wgt_tooltip
Function tooltip_fn

pub fn tooltip_fn(
    child: impl UiNode,
    tip: impl IntoVar<WidgetFn<TooltipArgs>>,
) -> impl UiNode

P Widget tooltip set as a widget function that is called every time the tooltip must be shown.

The tip widget function is used to instantiate a new tip widget when one needs to be shown, any widget can be used as tooltip, the recommended widget is the Tip! container, it provides the tooltip style.

This property can be configured by tooltip_anchor, tooltip_delay, tooltip_interval and tooltip_duration.

This tooltip only opens if the widget is enabled, see disabled_tooltip_fn for a tooltip that only shows when the widget is disabled.