Crate zng_wgt_tooltip

Tooltip widget, properties and nodes.

Crate

This crate is part of the zng project.

Structs

DefaultStyle

W Tip default style.

Tip

W A tooltip popup.

TooltipArgs

Arguments for tooltip widget functions.

Statics

ACCESS_TOOLTIP_ANCHOR_VAR

Position of the tip widget in relation to the anchor widget, when opened without cursor interaction.

ACCESS_TOOLTIP_DURATION_VAR

Maximum time a tooltip stays open, when opened without cursor interaction.

OPEN_TOOLTIP

Id of the current open tooltip.

STYLE_FN_VAR

Contextual style variable.

TOOLTIP_ANCHOR_VAR

Position of the tip widget in relation to the anchor widget, when opened with cursor interaction.

TOOLTIP_CONTEXT_CAPTURE_VAR

Tooltip context capture.

TOOLTIP_DELAY_VAR

Duration the cursor must be over the anchor widget before the tip widget is opened.

TOOLTIP_DURATION_VAR

Maximum time a tooltip stays open, when opened with cursor interaction.

TOOLTIP_INTERVAL_VAR

Maximum duration from the last time a tooltip was shown that a new tooltip opens instantly.

TOOLTIP_LAST_CLOSED

Tracks the instant the last tooltip was closed on the widget.

Functions

access_tooltip_anchor

P Set the position of the tip widgets opened for the widget or its descendants without cursor interaction.

access_tooltip_duration

P Sets the maximum duration a tooltip stays open on the widget or descendants when it is opened without cursor interaction.

disabled_tooltip

P Disabled widget tooltip.

disabled_tooltip_fn

P Disabled widget tooltip.

style_fn

P Extends or replaces the widget style.

tooltip

P Widget tooltip.

tooltip_anchor

P Set the position of the tip widgets opened for the widget or its descendants.

tooltip_context_capture

P Defines if the tooltip captures the build/instantiate context and sets it in the node context.

tooltip_delay

P Set the duration the cursor must be over the widget or its descendants before the tip widget is opened.

tooltip_duration

P Sets the maximum duration a tooltip stays open on the widget or descendants.

tooltip_fn

P Widget tooltip set as a widget function that is called every time the tooltip must be shown.

tooltip_interval

P Sets the maximum interval a second tooltip is opened instantly if a previous tip was just closed.